+++ src/main/java/org/logstash/netty/SslSimpleBuilder.java
@@ -5,7 +5,8 @@ import io.netty.handler.ssl.OpenSsl;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslContextBuilder;
 import io.netty.handler.ssl.SslHandler;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.logstash.beats.Server;
 
 import javax.net.ssl.SSLEngine;
@@ -29,7 +30,7 @@ public class SslSimpleBuilder {
         VERIFY_PEER,
         FORCE_PEER,
     }
-    private final static Logger logger = Logger.getLogger(SslSimpleBuilder.class);
+    private final static Logger logger = LogManager.getLogger(SslSimpleBuilder.class);
 
 
     private File sslKeyFile;
@@ -50,6 +51,7 @@ public class SslSimpleBuilder {
             "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256",
             "TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA384",
             "TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA384",
+            "TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256",
             "TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256"
     };
 
@@ -66,7 +68,7 @@ public class SslSimpleBuilder {
     }
 
     public SslSimpleBuilder setProtocols(String[] protocols) {
-        protocols = protocols;
+        this.protocols = protocols;
         return this;
     }
 
@@ -180,5 +182,13 @@ public class SslSimpleBuilder {
 
     private FileInputStream createFileInputStream(String filepath) throws FileNotFoundException {
         return new FileInputStream(filepath);
+    }
+
+    /**
+     * Get the supported protocols
+     * @return a defensive copy of the supported protocols
+     */
+    String[] getProtocols() {
+        return protocols.clone();
     }
 }